backend/requirements.txt
Django
djangorestframework
djangorestframework-csv
psycopg2-binary
pytz
